move pareto.py outside utils
